# Backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: ReForest API
  description: API para la gestión de la plataforma ReForest, que permite a los usuarios realizar donaciones para la reforestación y seguimiento de proyectos.
  version: "0.0.1-SNAPSHOT"
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
paths:
  /usuarios:
    get:
      tags:
        - Usuarios
      summary: Obtener lista de todos los usuarios
      description: Devuelve una lista con todos los usuarios registrados en el sistema.
      operationId: getAllUsuarios
      responses:
        '200':
          description: Lista de usuarios obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
              example:
                - id: 1
                  nombre: "Juan Pérez"
                  email: "juan.perez@example.com"
                  rol: "USER"
                - id: 2
                  nombre: "Ana López"
                  email: "ana.lopez@example.com"
                  rol: "ADMIN"
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Usuarios
      summary: Crear un nuevo usuario
      description: Registra un nuevo usuario en el sistema.
      operationId: createUsuario
      requestBody:
        description: Datos del nuevo usuario a crear.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '201':
          description: Usuario creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              example:
                id: 3
                nombre: "Carlos Gomez"
                email: "carlos.gomez@example.com"
                rol: "USER"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /usuarios/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener un usuario por su ID
      description: Devuelve los detalles de un usuario específico buscando por su ID.
      operationId: getUsuarioById
      parameters:
        - name: id
          in: path
          description: ID del usuario a obtener
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              example:
                id: 1
                nombre: "Juan Pérez"
                email: "juan.perez@example.com"
                rol: "USER"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Usuarios
      summary: Actualizar un usuario existente (completo)
      description: Modifica todos los datos de un usuario existente buscando por su ID.
      operationId: updateUsuario
      parameters:
        - name: id
          in: path
          description: ID del usuario a actualizar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Objeto completo del usuario con los datos actualizados.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '200':
          description: Usuario actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              example:
                id: 1
                nombre: "Juan Pérez Actualizado"
                email: "juan.perez.new@example.com"
                rol: "USER"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Usuarios
      summary: Actualizar parcialmente un usuario
      description: Actualiza uno o más campos de un usuario existente.
      operationId: patchUsuario
      parameters:
        - name: id
          in: path
          description: ID del usuario a actualizar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Un mapa de campos y valores a actualizar.
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                email: "nuevo.email@example.com"
                nombre: "Nombre Actualizado"
      responses:
        '200':
          description: Usuario actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              example:
                id: 1
                nombre: "Juan Pérez"
                email: "juan.perez.new@example.com"
                rol: "USER"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Usuarios
      summary: Eliminar un usuario por su ID
      description: Elimina un usuario del sistema de forma permanente.
      operationId: deleteUsuario
      parameters:
        - name: id
          in: path
          description: ID del usuario a eliminar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Usuario eliminado con éxito. No hay contenido en la respuesta.
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único del usuario.
          readOnly: true
        nombre:
          type: string
          description: Nombre completo del usuario.
          example: "Juan Pérez"
        email:
          type: string
          format: email
          description: Dirección de correo electrónico del usuario.
          example: "juan.perez@example.com"
        rol:
          type: string
          description: Rol del usuario en el sistema.
          enum: [USER, ADMIN]
          example: "USER"
    UsuarioInput:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre completo del usuario.
          example: "Juan Pérez"
        email:
          type: string
          format: email
          description: Dirección de correo electrónico del usuario.
          example: "juan.perez@example.com"
        contrasena:
          type: string
          format: password
          description: Contraseña del usuario (debe tener al menos 6 caracteres).
          example: "password123"
          writeOnly: true
        rol:
          type: string
          description: Rol del usuario en el sistema.
          enum: [USER, ADMIN]
          example: "USER"
      required:
        - nombre
        - email
        - contrasena
        - rol
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: Un código de error numérico HTTP.
          example: 404
        message:
          type: string
          description: Una descripción del error legible por humanos.
          example: "El recurso solicitado no fue encontrado."
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            message: "El recurso solicitado no se ha encontrado."
    BadRequest:
      description: Solicitud incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: "Los datos de la solicitud son inválidos."
    Conflict:
      description: Conflicto de datos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 409
            message: "Conflicto de datos. Probablemente un campo único (como el email) ya existe."
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: "Ha ocurrido un error inesperado en el servidor."